apps/api/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-environ==0.11.2
google-generativeai
npm